pyproject.toml
[build-system]
requires = ["setuptools>=68,<69", "wheel>=0.41,<0.43"]
build-backend = "setuptools.build_meta"

[project]
name = "forex-diffusion"
version = "0.1.0"
description = "MagicForex: FX intraday forecasting with latent diffusion + GUI"
readme = "README.md"
requires-python = ">=3.12,<3.13"
license = { text = "Proprietary" }
authors = [
  { name = "MagicForex Team" }
]
dependencies = [
  "torch>=2.3,<3.0",
  "pytorch-lightning>=2.3,<3.0",
  "einops>=0.7,<1.0",
  "numpy>=1.26,<3.0",
  "pandas>=2.2,<3.0",
  "pydantic>=2.6,<3.0",
  "fastapi>=0.111,<1.0",
  "uvicorn>=0.30,<1.0",
  "PyYAML>=6.0,<7.0",
  "httpx>=0.27,<1.0",
  "tenacity>=8.2,<9.0",
  "PySide6>=6.6,<7.0",
  "pyqtgraph>=0.13,<1.0",
  "alembic>=1.13,<2.0",
  "SQLAlchemy>=2.0,<3.0",
  "pyarrow>=15.0,<16.0",
  "tqdm>=4.66,<5.0"
]

[project.urls]
Homepage = "https://example.com/magicforex"
Repository = "https://example.com/magicforex/repo"

[project.scripts]
forex-gui = "forex_diffusion.ui.app:main"
forex-api = "forex_diffusion.inference.service:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["forex_diffusion*"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-q"
testpaths = ["tests"]


